+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add column continents (based on countries)
 - Retrieve missing publishers from Crossref
 - Add column indicating original research (as opposed to review, synthesis, opinion, etc)
+- Retrieve missing affiliations from OpenAlex (these are very messy currently)
 
 ## [1.0.0] - 2024-12-09